refactor: LottoController 위치 변경
