+++ src/main/java/io/suhan/lotto/Main.java
@@ -1,6 +1,6 @@
 package io.suhan.lotto;
 
-import io.suhan.lotto.model.lotto.LottoController;
+import io.suhan.lotto.controller.LottoController;
 
 public class Main {
     public static void main(String[] args) {
